fixx some

+++ src/main/java/ptit/repository/ManufactureDAO.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ptit.entity.Manufacturer;
 
-public interface ManufactureDAO extends JpaRepository<Manufacturer, Long> {
+public interface ManufactureDAO extends JpaRepository<Manufacturer, Integer> {
 	Manufacturer findByManufacturerID(int manufacturerID);
 	
 	List<Manufacturer> findAll();
